moved the encryption functions to go-util
